refactor - move config package to appropriate location and rename main package
