+++ src/main/java/cc/kertaskerja/config/DataConfig.java
@@ -1,4 +1,4 @@
-package cc.kertaskerja.realisasi_pemda_service.config;
+package cc.kertaskerja.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.r2dbc.config.EnableR2dbcAuditing;
