+++ src/test/java/cc/kertaskerja/RealisasiServiceTests.java
@@ -1,4 +1,4 @@
-package cc.kertaskerja.realisasi_pemda_service;
+package cc.kertaskerja;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Import;
 
 @Import(TestcontainersConfiguration.class)
 @SpringBootTest
-class RealisasiPemdaServiceApplicationTests {
+class RealisasiServiceTests {
 
 	@Test
 	void contextLoads() {
