+++ src/test/java/cc/kertaskerja/TestRealisasiServiceApplication.java
@@ -1,0 +1,11 @@
+package cc.kertaskerja;
+
+import org.springframework.boot.SpringApplication;
+
+public class TestRealisasiServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.from(RealisasiService::main).with(TestcontainersConfiguration.class).run(args);
+	}
+
+}
