+++ src/test/java/cc/kertaskerja/TestcontainersConfiguration.java
@@ -1,4 +1,4 @@
-package cc.kertaskerja.realisasi_pemda_service;
+package cc.kertaskerja;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
